Rename module from IntersectionNew to Intersection
