+++ frontend/src/modules/Intersection/DataProviderFramework/annotations/makeColorScaleAnnotation.ts
