+++ frontend/src/modules/Intersection/DataProviderFramework/boundingBoxes/makeGridBoundingBox.ts
