+++ frontend/src/modules/Intersection/DataProviderFramework/boundingBoxes/makeSeismicBoundingBox.ts
