+++ frontend/src/modules/Intersection/DataProviderFramework/boundingBoxes/makeSurfacesBoundingBox.ts
