+++ frontend/src/modules/Intersection/DataProviderFramework/customDataProviderImplementations/dataProviderTypes.ts
