+++ frontend/src/modules/Intersection/DataProviderFramework/customDataProviderImplementations/registerAllDataProviders.ts
