+++ frontend/src/modules/Intersection/DataProviderFramework/customDataProviderImplementations/utils.ts
