+++ frontend/src/modules/Intersection/DataProviderFramework/injectedDataType.ts
