+++ frontend/src/modules/Intersection/DataProviderFramework/utils.ts/colorScaleUtils.ts
