+++ frontend/src/modules/Intersection/DataProviderFramework/visualization/createSeismicLayerItemsMaker.ts
