+++ frontend/src/modules/Intersection/interfaces.ts
