+++ frontend/src/modules/Intersection/settings/atoms/baseAtoms.ts
@@ -1,8 +1,7 @@
 import { atom } from "jotai";
 
 import type { DataProviderManager } from "@modules/_shared/DataProviderFramework/framework/DataProviderManager/DataProviderManager";
-import { PreferredViewLayout } from "@modules/IntersectionNew/typesAndEnums";
-
+import { PreferredViewLayout } from "@modules/Intersection/typesAndEnums";
 
 export const userSelectedFieldIdentifierAtom = atom<string | null>(null);
 export const dataProviderManagerAtom = atom<DataProviderManager | null>(null);
