+++ frontend/src/modules/Intersection/settings/components/dataProviderManagerWrapper.tsx
@@ -32,8 +32,8 @@ import { GroupType } from "@modules/_shared/DataProviderFramework/groups/groupTy
 import type { Item, ItemGroup } from "@modules/_shared/DataProviderFramework/interfacesAndTypes/entities";
 import { Setting } from "@modules/_shared/DataProviderFramework/settings/settingsDefinitions";
 import { usePublishSubscribeTopicValue } from "@modules/_shared/utils/PublishSubscribeDelegate";
-import { CustomDataProviderType } from "@modules/IntersectionNew/DataProviderFramework/customDataProviderImplementations/dataProviderTypes";
-import { PreferredViewLayout } from "@modules/IntersectionNew/typesAndEnums";
+import { CustomDataProviderType } from "@modules/Intersection/DataProviderFramework/customDataProviderImplementations/dataProviderTypes";
+import { PreferredViewLayout } from "@modules/Intersection/typesAndEnums";
 
 import { preferredViewLayoutAtom } from "../atoms/baseAtoms";
