+++ frontend/src/modules/Intersection/settings/settings.tsx
