+++ frontend/src/modules/Intersection/typesAndEnums.ts
