+++ frontend/src/modules/Intersection/view/hooks/queryHooks.ts
