+++ frontend/src/modules/Intersection/view/hooks/useIntersectionReferenceSystem.ts
