+++ frontend/src/modules/Intersection/view/utils/boundingBoxUtils.ts
