+++ frontend/src/modules/Intersection/view/utils/createIntersectionReferenceSystem.ts
