+++ frontend/src/modules/Intersection/view/utils/createReferenceLines.tsx
