+++ frontend/src/modules/Intersection/view/utils/createWellboreLayerItems.tsx
