+++ frontend/src/modules/Intersection/view/view.tsx
